Add rebuildGrid() method
Implemented rebuildGrid(GridAttributes). It releases the MediaPlayers (with the new function release()), calls buildGrid(), sets the border size again and, for each item saved in items ArrayList, rebuildGrid() adds new views corresponding to the ones that were added before.
Such ArrayList is populated as each item is added. It uses Item class, also introduced in the version. Item class has Video, Image and Button especification subclasses. For now, these classes are only used to populate the ArrayList so they're private. Not sure if this the best aproach cost-wise, but it works.
Besides that, CollageView is not accepting any view as child anymore.